pyproject.toml
[tool.poetry]
name = "tbrandchatbot"
version = "0.1.0"
description = ""
authors = ["Sunho Lee <sunholee1217@sk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.2.8"
streamlit = "^1.40.1"
python-dotenv = "^1.0.1"
openai = "^1.54.5"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
chroma = "^0.2.0"
langchain-chroma = "^0.1.4"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
pypdf2 = "^3.0.1"
gspread = "^6.1.4"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
